.claude/skills/vision/pyproject.toml
[project]
name = "vision"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "pillow>=12.0.0",
    "markitdown",
    "zai-sdk",
]
